Split respawn toilet into trigger-and-consequence
Also move ExplodeConsequence to proper path
