+++ Assets/Scripts/Consequences/SetRespawnConsequence.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4e3e15cdb7f7cd048b27353839bd56a5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
