--- Assets/Scripts/TriggeredEffects/ExplodeConsequence.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 27766b92662dd1844b3677d56635c535
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences:
-  - ExplosionPrefab: {fileID: 1290715649332758, guid: 587eaecaa621be74486f351c78b4f0a8, type: 3}
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
